Refactor connector, wrap errors
